refactor: [#4] Options 분리
