+++ src/constants/options/ageOptions.ts
@@ -1,0 +1,12 @@
+const ageOptions = [
+  { label: '10대', value: '10' },
+  { label: '20대', value: '20' },
+  { label: '30대', value: '30' },
+  { label: '40대', value: '40' },
+  { label: '50대', value: '50' },
+  { label: '60대', value: '60' },
+  { label: '70대', value: '70' },
+  { label: '80대', value: '80' },
+]
+
+export default ageOptions
