+++ src/constants/options/genderOptions.ts
@@ -1,0 +1,6 @@
+const genderOptions = [
+  { label: '남성', value: 'male' },
+  { label: '여성', value: 'female' },
+]
+
+export default genderOptions
